undo

+++ run_trainer_tpu.py
@@ -37,10 +37,10 @@ def main():
     with torch.no_grad():
         for module in model.modules():
             if isinstance(module, SharedMatrix):
-                module.matrix.data = module.matrix.bfloat16()
+                module.matrix.data = module.matrix.half()
             if isinstance(module, AdaptedLinear):
-                module.adapter_first.data = module.adapter_first.bfloat16()
-                module.adapter_second.data = module.adapter_second.bfloat16()
+                module.adapter_first.data = module.adapter_first.half()
+                module.adapter_second.data = module.adapter_second.half()
 
     # BEGIN init TPU
     assert trainer_args.do_train and not trainer_args.do_eval
@@ -56,8 +56,8 @@ def main():
         start=True,
     )
     assert model is task.model and model is tpu_manager._synchronizer.master_model
-    assert any(param.dtype == torch.bfloat16 for param in optimizer.state_averager.main_parameters)
-    assert not any(param.dtype == torch.bfloat16 for group in optimizer.state_averager.optimizer.param_groups for param in group["params"])
+    assert any(param.dtype == torch.float16 for param in optimizer.state_averager.main_parameters)
+    assert not any(param.dtype == torch.float16 for group in optimizer.state_averager.optimizer.param_groups for param in group["params"])
 
 
     # warmup tpus
